refactor(frontend): extract Chips + MRow + Chk (9/N)
Three small presentational components moved out:
- components/Chips.jsx   — multi-select filter chips (with pin-to-top)
- components/MRow.jsx    — bundles MRow + Chk (both Monitor-tab helpers)

App.jsx -52 LOC. Build clean.
